requirements-api.txt
# Minimal dependencies for API inference only

# API
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.12.4

# Core ML (inference only - CPU optimized)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
transformers==4.46.0
peft==0.13.0
safetensors==0.7.0

# MLOps (client only)
mlflow-skinny==2.16.0

# Monitoring
python-json-logger>=2.0.7

# Utilities
PyYAML==6.0.3
python-dotenv==1.2.1
